添加注解，更改token主键数据类型，完善标签查询方法

+++ src/main/java/com/syt/blog/repository/RefreshTokenRepository.java
@@ -9,5 +9,5 @@ public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long
 
     Optional<RefreshToken> findByToken(String token);
 
-    void deleteByUserIdAndToken(Long userId, String token);
+    void deleteByUserIdAndToken(Integer userId, String token);
 }
