+++ src/main/java/com/syt/blog/service/RefreshTokenService.java
@@ -4,7 +4,7 @@ import com.syt.blog.entity.RefreshToken;
 
 public interface RefreshTokenService {
 
-    RefreshToken saveRefreshToken(Long userId, String token, String deviceId, String userAgent, String ip);
+    RefreshToken saveRefreshToken(Integer userId, String token, String deviceId, String userAgent, String ip);
 
     RefreshToken verifyRefreshToken(String token);
